+++ resource/javascript/setting.js
@@ -1,3 +1,6 @@
+load_script("../javascript/component/search_box.js");
+load_script("../javascript/component/warning_message.js");
+
 new Konami(function () {
     const body_element = document.body;
     body_element.style.transition = "2s";
@@ -42,22 +45,13 @@ async function set_to_current_setting() {
     document.getElementById("speed").value = document.getElementById("speed_value").innerHTML = await eel.read_setting("speed")();
     document.getElementById("volume").value = document.getElementById("volume_value").innerHTML = await eel.read_setting("volume")();
 
-    if (await eel.read_setting("bold_text")() == "True") {
-        document.getElementById("bold_text").checked = true;
-    }
-    if (await eel.read_setting("bigger_text")() == "True") {
-        document.getElementById("bigger_text").checked = true;
-    }
-    if (await eel.read_setting("continuous_speech_recognition")() == "True") {
-        document.getElementById("continuous_speech_recognition").checked = true;
-    }
-    if (await eel.read_flag("add_readable_text_setting")()) {
-        document.getElementById("readable_text_setting").style.display = "block";
-    }
+    document.getElementById("bold_text").checked = await eel.read_setting("bold_text")();
+    document.getElementById("bigger_text").checked = await eel.read_setting("bigger_text")();
+    document.getElementById("continuous_speech_recognition").checked = await eel.read_setting("continuous_speech_recognition")();
+    document.getElementById("show_tips").checked = await eel.read_setting("show_tips")();
 
-    if (await eel.read_flag("enable_improved_custom_theme_editor")()) {
-        document.getElementById("edit_theme").href = "improved_set_custom_css_theme.html";
-    }
+    if (await eel.read_flag("add_readable_text_setting")()) document.getElementById("readable_text_setting").style.display = "block";
+    if (await eel.read_flag("show_tips_setting")()) document.getElementById("tips_setting").style.display = "block";
 }
 
 set_to_current_setting();
@@ -121,15 +115,34 @@ document.querySelectorAll("mwc-radio[name='search_engine_list']").forEach((eleme
 });
 
 document.getElementById("check_update").addEventListener("click", async function () {
+    function sanitize(string) {
+        const sanitize_table = {
+            "&": '&amp;',
+            "'": '&#x27;',
+            "`": '&#x60;',
+            "\"": '&quot;',
+            "<": '&lt;',
+            ">": '&gt;',
+        };
+        Object.keys(sanitize_table).forEach((key) => {
+            string.replaceAll(key, sanitize_table[key]);
+        });
+        return string;
+    }
+
     document.getElementById("update_detail").innerHTML = "アップデートを確認中です...";
-    const response = await fetch("../information.txt");
-    const information_data = await response.text();
-    const version_information = information_data.match(/Version:.*/)[0].replace("Version:", "");
-    const release_data = version_information.indexOf("dev") === -1 ? await eel.get_release("stable")() : await eel.get_release("develop")();
+    const response = await fetch("../information.json");
+    const information_data = await response.json();
+    const version_information = information_data.Version;
+    const release_data = await eel.get_release(information_data.Channel)();
 
     const latest_version = compare_and_return_latest_version_num(version_information, release_data[0]);
-    const update_detail = latest_version === version_information ? "利用可能なアップデートはありません。最新のバージョンを使用中です。" : `<h1>${release_data[0]}</h1>${release_data[1]}<br><a href='https://github.com/Robot-Inventor/ORIZIN-Agent-HTML/releases' target='_blank' rel='noreferrer noopener' class='ripple_effect'>ダウンロード<i class='material_icon'>open_in_new</i></a>`;
-    document.getElementById("update_detail").innerHTML = update_detail;
+    const update_detail = document.getElementById("update_detail");
+    if (latest_version === version_information) {
+        update_detail.innerHTML = "利用可能なアップデートはありません。最新のバージョンを使用中です。";
+    } else {
+        update_detail.innerHTML = `<h1>${release_data[0]}</h1>${release_data[1]}<br><a href='https://github.com/Robot-Inventor/ORIZIN-Agent-HTML/releases' target='_blank' rel='noreferrer noopener' class='ripple_effect'>ダウンロード<i class='material_icon'>open_in_new</i></a>`;
+    }
 });
 
 document.getElementById("reset_setting_button").addEventListener("click", () => {
@@ -156,8 +169,11 @@ document.getElementById("factory_reset_button").addEventListener("click", () => 
     });
 });
 
-const search_box = document.getElementById("search_box");
-search_box.init("div.fill_panel section");
+window.addEventListener("load", () => {
+    const search_box = document.getElementById("search_box");
+    search_box.init("div.fill_panel section");
+});
+
 Mousetrap.bind({
     "/": () => {
         search_box.focus();